moved def of interfaces
