--- ChoreBuddies.Backend/Features/Notifications/Email/IEmailService.cs
@@ -1,11 +1,0 @@
-﻿using ChoreBuddies.Backend.Domain;
-
-namespace ChoreBuddies.Backend.Features.Notifications.Email;
-
-public interface IEmailService
-{
-    public Task<string> SendRegisterConfirmationNotificationAsync(
-        AppUser recipientEmail,
-        string recipientName,
-        CancellationToken cancellationToken = default);
-}
